# tunis-event-app-backend/src/modules/auth/dto/login.dto.ts
import { IsEmail, IsString } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class LoginDto {
  @ApiProperty({ example: 'user@example.com' })
  @IsEmail()
  email!: string;
  @ApiProperty({ example: 'securePassword123' })
  @IsString()
  password!: string;
}