# tunis-event-app-backend/src/modules/auth/dto/register.dto.ts
import { IsEmail, IsString, IsEnum } from 'class-validator';
import { Role } from '@prisma/client';
import { ApiProperty } from '@nestjs/swagger';

export class RegisterDto {
  @ApiProperty({ example: 'user@example.com' })
  @IsEmail()
  email!: string;

  @ApiProperty({ example: 'securePassword123' })
  @IsString()
  password!: string;

  @ApiProperty({ example: 'John Doe' })
  @IsString()
  name!: string;

  @ApiProperty({ enum: Role })
  @IsEnum(['SUPERADMIN', 'ORGANISATEUR', 'PARTICIPANT'], {
    message: 'Invalid role',
  })
  role!: 'SUPERADMIN' | 'ORGANISATEUR' | 'PARTICIPANT';
}