# tunis-event-app/openapi-backend.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/register-organisateur":{"post":{"operationId":"AuthController_registerOrganisateur","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterOrganisateurDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/organization":{"post":{"operationId":"OrganizationController_create","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["organization"]}},"/profile":{"get":{"operationId":"ProfileController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Profile"]}}},"info":{"title":"Tunis Events API","description":"API pour les événements avec authentification","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"securePassword123"},"name":{"type":"string","example":"John Doe"},"role":{"type":"string","enum":["SUPERADMIN","ORGANISATEUR","PARTICIPANT"]}},"required":["email","password","name","role"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"securePassword123"}},"required":["email","password"]},"RegisterOrganisateurDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"securePassword123"},"name":{"type":"string","example":"John Doe"},"role":{"type":"string","enum":["SUPERADMIN","ORGANISATEUR","PARTICIPANT"]}},"required":["email","password","name","role"]}}}}